requirements.txt
flask==2.3.2
numpy==1.26.1
opencv-python-headless==4.8.1.78
pandas==2.1.0
openpyxl==3.1.3
supabase==1.0.0
paho-mqtt==1.6.1
